chaser - move generic CVR git log scrapper to common.py
